+++ salesmanago/salesmanago.php
@@ -2,9 +2,9 @@
 /**
  * Plugin Name: SALESmanago & Leadoo
  * Plugin URI:  https://www.salesmanago.com/?utm_source=integration&utm_medium=WORDPRESS&utm_content=marketplace
- * Description: SALESmanago Marketing Automation integration for WordPress, WooCommerce, Contact Form 7, Gravity Forms
- * Version:     3.9.1
- * Tested up to: 6.8.3
+ * Description: SALESmanago Marketing Automation integration for WordPress, WooCommerce, Contact Form 7, Gravity Forms, Tier Pricing
+ * Version:     3.10.0
+ * Tested up to: 6.9.1
  * Requires PHP: 7.4
  * Author:      SALESmanago
  * Author URI:  https://www.salesmanago.com/?utm_source=integration&utm_medium=WORDPRESS&utm_content=marketplace